requirements.txt
azure-ai-vision-imageanalysis>=1.0.0
azure-core>=1.34.0
dotenv>=0.9.9
pillow>=11.2.1
piexif
pillow-heif